Polish/last minute bugfixes and polish

+++ Assets/Scripts/Upgrades/UpgradeStats/AttackUpgradeStats.cs
@@ -3,7 +3,7 @@
 [CreateAssetMenu(fileName = "Upgrade Stats", menuName = "ScriptableObjects/UpgradeStats/DamageUpgrade", order = 1)]
 public class AttackUpgradeStats : UpgradeStats
 {
-    public int extraDamage;
+    public float extraDamage;
     
     public override void Apply(StatsManager manager)
     {
