+++ Assets/Scripts/Weapons/WeaponData.cs
@@ -6,8 +6,8 @@ public abstract class WeaponData : ScriptableObject
 {
     public float cooldown;
     public float attackLifetime;
-    public int damage;
-    public int knockback;
+    public float damage;
+    public float knockback;
     public int weaponDataID = -1;
 
     public abstract void Attack(GameObject player);
